Simple babel for es2016 syntax
