+++ package.json
@@ -7,7 +7,7 @@
   "description": "Texts my sister every so often to wish her a happy birthday",
   "main": "main.js",
   "scripts": {
-    "start": "node main.js"
+    "start": "nodemon main.js --exec babel-node"
   },
   "keywords": [
     "birthday",
@@ -16,7 +16,14 @@
   "author": "Ari Sweedler",
   "license": "MIT",
   "dependencies": {
+    "@babel/cli": "^7.13.16",
+    "@babel/core": "^7.14.0",
+    "@babel/node": "^7.13.13",
+    "@babel/preset-env": "^7.14.0",
     "express": "^4.17.1",
     "twilio": "^3.63.0"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.7"
   }
 }
